chore: remove version sync script and simplify version management
