--- crates/loro-internal/VERSION
@@ -1,1 +1,0 @@
-1.4.2
